# test/crimson/seastore/test_omap_manager: redesign test variable sizes