@tchaikov 2025-07-16
jenkins test make check